Move slice building into SliceRegistrar
This seems like a more appropriate place for it than on  Slice itself, since it'd otherwise make that method available on each slice subclass used within the app.
